+++ src/Descript/Ast/Flat.hs
@@ -1,0 +1,2 @@
+-- | Flat parsing of AST data.
+{-# OPTIONS_GHC -F -pgmF autoexporter #-}