+++ src/Descript/Compile.hs
@@ -11,16 +11,17 @@ import Descript.Misc
 import Descript.Plugin
 
 import Control.Monad
-import Control.Monad.IO.Class
 import qualified Data.Text.IO as T
 
 -- | Read the Descript source from the first path, compile it, and move the executable to the second path.
 compile :: FilePath -> FilePath -> SessionRes ()
 compile input output
+  | input == output = mkFail $ mkOverlapInOutError StageReadArgs
+  | otherwise
   = ( (`T.compile` output)
   <=< T.translate
   <=< C.parse
   <=< ResultT . pure . S.parse
   <=< ResultT . pure . L.parse
-  <=< liftIO . T.readFile
+  <=< liftIOAndCatch StageReadInput . T.readFile
     ) input